consumer/requirements.txt
kafka-python
rdflib
requests
